global config ngrx in seprate files
