+++ components/automate-ui/src/app/entities/destinations/destination.actions.ts
@@ -1,7 +1,7 @@
 import { HttpErrorResponse } from '@angular/common/http';
 import { Action } from '@ngrx/store';
 
-import { Destination, EnableDestination, GlobalConfig } from './destination.model';
+import { Destination, EnableDestination } from './destination.model';
 
 export enum DestinationActionTypes {
   GET_ALL                        = 'DESTINATION::GET_ALL',
@@ -24,19 +24,14 @@ export enum DestinationActionTypes {
   SEND_TEST_FAILURE              = 'DESTINATION::SEND_TEST::FAILURE',
   ENABLE_DISABLE                 = 'DESTINATION::ENABLE_DISABLE',
   ENABLE_DISABLE_SUCCESS         = 'DESTINATION::ENABLE_DISABLE::SUCCESS',
-  ENABLE_DISABLE_FAILURE         = 'DESTINATION::ENABLE_DISABLE::FAILURE',
-  GLOBAL_CONFIG                  = 'DESTINATION::GLOBAL_CONFIG',
-  GLOBAL_CONFIG_SUCCESS          = 'DESTINATION::GLOBAL_CONFIG::SUCCESS',
-  GLOBAL_CONFIG_FAILURE          = 'DESTINATION::GLOBAL_CONFIG::FAILURE'
+  ENABLE_DISABLE_FAILURE         = 'DESTINATION::ENABLE_DISABLE::FAILURE'
 }
 
 
 export interface DestinationSuccessPayload {
   destination: Destination;
 }
-export interface  GlobalDataFeedConfigSuccessPayload {
-  config: GlobalConfig;
-}
+
 
 export interface GetDestinationsSuccessPayload {
   destinations: Destination[];
@@ -167,21 +162,6 @@ export class EnableDisableDestinationFailure implements Action {
 
   constructor(public payload: HttpErrorResponse) { }
 }
-export class GlobalDataFeedConfig implements Action {
-  readonly type = DestinationActionTypes.GLOBAL_CONFIG;
-
-  constructor(public payload) { }
-}
-export class GlobalDataFeedConfigSuccess implements Action {
-  readonly type = DestinationActionTypes.GLOBAL_CONFIG_SUCCESS;
-
-  constructor(public payload) { }
-}
-export class GlobalDataFeedConfigFailure implements Action {
-  readonly type = DestinationActionTypes.GLOBAL_CONFIG_FAILURE;
-
-  constructor(public payload: HttpErrorResponse) { }
-}
 
 export type DestinationActions =
   | GetDestinations
@@ -204,7 +184,4 @@ export type DestinationActions =
   | TestDestinationFailure
   | EnableDisableDestination
   | EnableDisableDestinationSuccess
-  | EnableDisableDestinationFailure
-  | GlobalDataFeedConfig
-  | GlobalDataFeedConfigSuccess
-  | GlobalDataFeedConfigFailure;
+  | EnableDisableDestinationFailure;
