+++ components/automate-ui/src/app/entities/destinations/destination.effects.ts
@@ -31,9 +31,7 @@ import {
   TestDestinationFailure,
   EnableDisableDestination,
   EnableDisableDestinationSuccess,
-  EnableDisableDestinationFailure,
-  GlobalDataFeedConfigSuccess,
-  GlobalDataFeedConfigFailure
+  EnableDisableDestinationFailure
 } from './destination.actions';
 
 import {
@@ -194,54 +192,36 @@ export class DestinationEffects {
 
   testDestinationFailure$ = createEffect(() =>
     this.actions$.pipe(
-    ofType(DestinationActionTypes.SEND_TEST_FAILURE),
-    map(({ payload }: TestDestinationFailure) => {
-      return new CreateNotification({
-        type: Type.error,
-        message: `Unable to connect to data feed ${payload.name}.`
-      });
+      ofType(DestinationActionTypes.SEND_TEST_FAILURE),
+      map(({ payload }: TestDestinationFailure) => {
+        return new CreateNotification({
+          type: Type.error,
+          message: `Unable to connect to data feed ${payload.name}.`
+        });
     })));
 
     enableDisableDestination$ = createEffect(() =>
-    this.actions$.pipe(
-    ofType(DestinationActionTypes.ENABLE_DISABLE),
-    mergeMap(( {payload:  {enableDisable} }: EnableDisableDestination) =>
-      this.requests.enableDestinations(enableDisable).pipe(
-        map((resp: DestinationSuccessPayload) => new EnableDisableDestinationSuccess(resp)),
-        catchError((error: HttpErrorResponse) =>
-          observableOf(new EnableDisableDestinationFailure(error)))))));
+      this.actions$.pipe(
+        ofType(DestinationActionTypes.ENABLE_DISABLE),
+        mergeMap(( {payload:  {enableDisable} }: EnableDisableDestination) =>
+          this.requests.enableDestinations(enableDisable).pipe(
+            map((resp: DestinationSuccessPayload) => new EnableDisableDestinationSuccess(resp)),
+            catchError((error: HttpErrorResponse) =>
+              observableOf(new EnableDisableDestinationFailure(error)))))));
 
     enableDisableDestinationFailure$ = createEffect(() =>
-    this.actions$.pipe(
-    ofType(DestinationActionTypes.ENABLE_DISABLE_FAILURE),
-    map(({ payload  }: EnableDisableDestinationFailure) => new CreateNotification({
-      type: Type.info,
-      message: `Could not enable or Disable: error ${payload.error}.`
-    }))));
+      this.actions$.pipe(
+        ofType(DestinationActionTypes.ENABLE_DISABLE_FAILURE),
+        map(({ payload  }: EnableDisableDestinationFailure) => new CreateNotification({
+          type: Type.info,
+          message: `Could not enable or Disable: error ${payload.error}.`
+        }))));
 
     enableDisableDestinationSuccess$ = createEffect(() =>
-    this.actions$.pipe(
-      ofType(DestinationActionTypes.ENABLE_DISABLE_SUCCESS),
-      map(({ payload  }: EnableDisableDestinationSuccess) => new CreateNotification({
-      type: Type.info,
-      message: `Destination is ${payload.enable ? 'Enabled' : 'Disabled'}.`
-    }))));
-
-    globalDataFeedConfig$ = createEffect(() =>
-    this.actions$.pipe(
-    ofType(DestinationActionTypes.GLOBAL_CONFIG),
-    mergeMap(() =>
-      this.requests.globalDataFeedConfig().pipe(
-        map((resp: GlobalDataFeedConfigSuccess) => new GlobalDataFeedConfigSuccess(resp)),
-        catchError((error: HttpErrorResponse) =>
-          observableOf(new GetDestinationsFailure(error)))))));
-
-    globalDataFeedConfigFailure$ = createEffect(() =>
-    this.actions$.pipe(
-    ofType(DestinationActionTypes.GLOBAL_CONFIG_FAILURE),
-    map(({ payload  }: GlobalDataFeedConfigFailure) => new CreateNotification({
-      type: Type.error,
-      message: `Error while fetching data-feed config: error ${payload.error}.`
-    }))));
-
+      this.actions$.pipe(
+        ofType(DestinationActionTypes.ENABLE_DISABLE_SUCCESS),
+        map(({ payload  }: EnableDisableDestinationSuccess) => new CreateNotification({
+          type: Type.info,
+          message: `Destination is ${payload.enable ? 'Enabled' : 'Disabled'}.`
+      }))));
 }
