+++ components/automate-ui/src/app/entities/global-config/destination-config.model.ts
@@ -1,0 +1,8 @@
+export interface GlobalConfig {
+  feed_interval: string;
+  node_batch_size: string;
+  updated_nodes_only: boolean;
+  disable_cidr_filter: boolean;
+  cidr_filter: string[];
+  accepted_status_codes: number[];
+}
